Delete Entities due data model refactor
