+++ app/src/main/java/ru/jjba/jr2/data/db/dao/WordDao.kt
@@ -16,9 +16,6 @@ abstract class WordDao {
     @Query("SELECT * FROM Word")
     abstract fun getAll(): Flowable<List<Word>>
 
-    @Query("SELECT * FROM Word WHERE basicInterpretation LIKE :query OR wordJp LIKE :query")
-    abstract fun find(query: String): Flowable<List<Word>>
-
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     abstract fun insert(word: Word): Long
