--- app/src/main/java/ru/jjba/jr2/domain/entity/Example.kt
@@ -1,16 +1,0 @@
-package ru.jjba.jr2.domain.entity
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity
-class Example(@PrimaryKey
-              val id: String,
-              val exampleJp: String,
-              val exampleFurigana: String,
-              val exampleTranslate: String,
-        // TODO: Использовать @Embedded(prefix = "interpretation_")
-              val interp: String
-) {
-    override fun toString() = "$id $exampleJp $exampleFurigana $exampleTranslate"
-}
