--- app/src/main/java/ru/jjba/jr2/domain/entity/Interpretation.kt
@@ -1,16 +1,0 @@
-package ru.jjba.jr2.domain.entity
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity
-class Interpretation(
-        @PrimaryKey(autoGenerate = true)
-        var id: Long,
-        var interp: String,
-        var pos: String,
-        //@Embedded(prefix = "word_")
-        var word: Long
-) {
-    override fun toString() = "$id $pos $interp"
-}
