--- app/src/main/java/ru/jjba/jr2/domain/interactor/KanjiInteractor.kt
@@ -1,2 +1,0 @@
-package ru.jjba.jr2.domain.interactor
-
