+++ app/src/main/java/ru/jjba/jr2/presentation/ui/vocab/word/group/WordGroupListAdapter.kt
@@ -1,0 +1,4 @@
+package ru.jjba.jr2.presentation.ui.vocab.word.group
+
+class WordGroupListAdapter {
+}
